PPO/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pytesseract>=0.3.10
Pillow>=10.0.0
python-dotenv>=1.0.0
google-generativeai
